requirements.txt
typer==0.12.3
pyyaml==6.0.1
watchdog==4.0.1
pytest==7.4.4
